Refactor project dashboard into reusable components
