--- src/App.css
@@ -1,467 +1,0 @@
-:root {
-  font-family: 'Inter', system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
-  line-height: 1.5;
-  font-weight: 400;
-  color: #0f172a;
-  background-color: #f5f7fb;
-  --color-success: #34d399;
-  --color-warning: #fbbf24;
-  --color-danger: #f87171;
-  --color-info: #60a5fa;
-}
-
-* {
-  box-sizing: border-box;
-}
-
-body {
-  margin: 0;
-  background-color: #f5f7fb;
-}
-
-.dashboard {
-  min-height: 100vh;
-  padding: 2.5rem;
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-}
-
-.dashboard-header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 1rem;
-}
-
-.dashboard-header__title {
-  display: flex;
-  align-items: center;
-  gap: 1.5rem;
-}
-
-.dashboard-header h1 {
-  margin: 0;
-  font-size: 2rem;
-  font-weight: 700;
-}
-
-.dashboard-header select {
-  border: 1px solid #e2e8f0;
-  border-radius: 12px;
-  padding: 0.5rem 1rem;
-  font-size: 1rem;
-  background: #fff;
-  color: inherit;
-}
-
-.header-button {
-  border: none;
-  border-radius: 999px;
-  padding: 0.65rem 1.4rem;
-  background: #2563eb;
-  color: #fff;
-  font-weight: 600;
-  display: flex;
-  align-items: center;
-  gap: 0.4rem;
-  cursor: pointer;
-  transition: background 150ms ease-in-out, transform 150ms ease-in-out;
-}
-
-.header-button:hover {
-  background: #1d4ed8;
-  transform: translateY(-1px);
-}
-
-.dashboard__layout {
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-}
-
-.stats-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(220px, 1fr));
-  gap: 1rem;
-}
-
-.stats-card {
-  background: #fff;
-  border-radius: 20px;
-  padding: 1.25rem;
-  display: flex;
-  gap: 1rem;
-  align-items: center;
-  box-shadow: 0 10px 30px rgba(15, 23, 42, 0.08);
-}
-
-.stats-card__icon {
-  width: 48px;
-  height: 48px;
-  border-radius: 16px;
-  display: grid;
-  place-items: center;
-  font-size: 1.5rem;
-  background: #eff4ff;
-}
-
-.stats-card__content {
-  display: flex;
-  flex-direction: column;
-  gap: 0.2rem;
-}
-
-.stats-card__title {
-  margin: 0;
-  font-size: 0.95rem;
-  color: #64748b;
-  font-weight: 600;
-}
-
-.stats-card__value {
-  margin: 0;
-  font-size: 1.8rem;
-  font-weight: 700;
-}
-
-.stats-card__subtitle {
-  margin: 0;
-  font-size: 0.85rem;
-  color: #94a3b8;
-}
-
-.stats-card--default .stats-card__icon {
-  background: rgba(99, 102, 241, 0.1);
-}
-
-.stats-card--success .stats-card__icon {
-  background: rgba(52, 211, 153, 0.15);
-}
-
-.stats-card--danger .stats-card__icon {
-  background: rgba(248, 113, 113, 0.15);
-}
-
-.stats-card--warning .stats-card__icon {
-  background: rgba(251, 191, 36, 0.15);
-}
-
-.stats-card--info .stats-card__icon {
-  background: rgba(96, 165, 250, 0.15);
-}
-
-.dashboard__content {
-  display: grid;
-  grid-template-columns: minmax(0, 2fr) minmax(0, 1fr);
-  gap: 1.5rem;
-  align-items: start;
-}
-
-.dashboard__main-panels {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(260px, 1fr));
-  gap: 1.5rem;
-}
-
-.panel {
-  background: #fff;
-  border-radius: 20px;
-  padding: 1.5rem;
-  box-shadow: 0 10px 30px rgba(15, 23, 42, 0.08);
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-}
-
-.panel__header h2 {
-  margin: 0;
-  font-size: 1.2rem;
-  font-weight: 700;
-  color: #0f172a;
-}
-
-.sprint-status {
-  display: flex;
-  gap: 2rem;
-}
-
-.sprint-status__chart {
-  width: 160px;
-  height: 160px;
-  position: relative;
-}
-
-.sprint-status__donut {
-  width: 100%;
-  height: 100%;
-  border-radius: 50%;
-  background: conic-gradient(
-    var(--color-success) 0deg 226.8deg,
-    var(--color-info) 226.8deg 270deg,
-    var(--color-warning) 270deg 360deg
-  );
-  display: grid;
-  place-items: center;
-}
-
-.sprint-status__donut-hole {
-  width: 70px;
-  height: 70px;
-  border-radius: 50%;
-  background: #fff;
-  box-shadow: inset 0 0 8px rgba(15, 23, 42, 0.08);
-}
-
-.sprint-status__legend {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  justify-content: center;
-}
-
-.legend-dot {
-  width: 12px;
-  height: 12px;
-  border-radius: 50%;
-  display: inline-block;
-  margin-right: 0.75rem;
-}
-
-.legend-label {
-  margin: 0;
-  font-weight: 600;
-}
-
-.legend-value {
-  margin: 0;
-  font-size: 0.85rem;
-  color: #64748b;
-}
-
-.task-priority {
-  display: flex;
-  gap: 1.5rem;
-  align-items: flex-end;
-}
-
-.task-priority__chart {
-  flex: 1;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 1rem;
-  align-items: end;
-  height: 220px;
-  padding: 0 0.5rem;
-}
-
-.task-priority__column {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 0.75rem;
-}
-
-.task-priority__bars {
-  display: flex;
-  flex-direction: column;
-  width: 60%;
-  gap: 0.4rem;
-  justify-content: flex-end;
-}
-
-.task-priority__bar {
-  width: 100%;
-  border-radius: 999px;
-}
-
-.task-priority__label {
-  margin: 0;
-  font-weight: 600;
-}
-
-.task-priority__legend {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-  font-size: 0.9rem;
-  color: #475569;
-}
-
-.team-progress {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.team-progress__item {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.team-progress__info {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.team-progress__name {
-  margin: 0;
-  font-weight: 600;
-}
-
-.team-progress__value {
-  margin: 0;
-  color: #64748b;
-  font-size: 0.9rem;
-}
-
-.team-progress__bar {
-  background: #f1f5f9;
-  border-radius: 999px;
-  height: 10px;
-  overflow: hidden;
-}
-
-.team-progress__fill {
-  height: 100%;
-  border-radius: inherit;
-  background: linear-gradient(90deg, #2563eb, #60a5fa);
-}
-
-.assistant-panel {
-  height: 100%;
-}
-
-.assistant-panel__title {
-  display: flex;
-  align-items: center;
-  gap: 0.75rem;
-}
-
-.assistant-panel__badge {
-  width: 36px;
-  height: 36px;
-  border-radius: 12px;
-  display: grid;
-  place-items: center;
-  background: rgba(37, 99, 235, 0.1);
-  font-size: 1.5rem;
-}
-
-.assistant-panel__conversation {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.assistant-message,
-.user-message {
-  display: flex;
-}
-
-.assistant-message {
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 0.4rem;
-}
-
-.user-message {
-  justify-content: flex-end;
-}
-
-.assistant-message__meta {
-  margin: 0;
-  font-size: 0.85rem;
-  color: #64748b;
-}
-
-.assistant-message__bubble {
-  margin: 0;
-  padding: 0.9rem 1rem;
-  background: rgba(37, 99, 235, 0.08);
-  border-radius: 16px;
-  color: #1e3a8a;
-}
-
-.user-message__bubble {
-  margin: 0;
-  align-self: flex-end;
-  padding: 0.9rem 1rem;
-  background: #2563eb;
-  color: #fff;
-  border-radius: 16px 16px 4px 16px;
-}
-
-.assistant-panel__input {
-  margin-top: auto;
-  display: flex;
-  gap: 0.75rem;
-}
-
-.assistant-panel__input input {
-  flex: 1;
-  border-radius: 999px;
-  border: 1px solid #e2e8f0;
-  padding: 0.8rem 1.2rem;
-  font-size: 0.95rem;
-}
-
-.assistant-panel__input button {
-  border-radius: 999px;
-  border: none;
-  padding: 0.8rem 1.5rem;
-  font-weight: 600;
-  background: #2563eb;
-  color: #fff;
-  cursor: pointer;
-}
-
-.visually-hidden {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  padding: 0;
-  margin: -1px;
-  overflow: hidden;
-  clip: rect(0, 0, 0, 0);
-  border: 0;
-}
-
-@media (max-width: 1024px) {
-  .dashboard__content {
-    grid-template-columns: 1fr;
-  }
-
-  .assistant-panel {
-    order: -1;
-  }
-}
-
-@media (max-width: 768px) {
-  .dashboard {
-    padding: 1.5rem;
-  }
-
-  .dashboard-header {
-    flex-direction: column;
-    align-items: flex-start;
-  }
-
-  .dashboard-header__title {
-    width: 100%;
-    justify-content: space-between;
-  }
-
-  .header-button {
-    width: 100%;
-    justify-content: center;
-  }
-}
