+++ src/components/SprintStatus.css
@@ -1,0 +1,61 @@
+.sprint-status {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 2rem;
+}
+
+.sprint-status__chart {
+  width: 160px;
+  height: 160px;
+  position: relative;
+}
+
+.sprint-status__donut {
+  width: 100%;
+  height: 100%;
+  border-radius: 50%;
+  background: conic-gradient(
+    var(--color-success) 0deg 226.8deg,
+    var(--color-warning) 226.8deg 316.8deg,
+    var(--color-info) 316.8deg 360deg
+  );
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.sprint-status__donut-hole {
+  width: 92px;
+  height: 92px;
+  border-radius: 50%;
+  background: #fff;
+  box-shadow: inset 0 0 0 4px rgba(15, 23, 42, 0.05);
+}
+
+.sprint-status__legend {
+  list-style: none;
+  margin: 0;
+  padding: 0;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.sprint-status__legend li {
+  display: flex;
+  align-items: center;
+  gap: 0.75rem;
+}
+
+.legend-label {
+  margin: 0;
+  font-size: 0.95rem;
+  font-weight: 600;
+  color: #0f172a;
+}
+
+.legend-value {
+  margin: 0;
+  font-size: 0.85rem;
+  color: #64748b;
+}
