+++ src/components/SprintStatus.tsx
@@ -1,4 +1,5 @@
 import type { FC } from 'react';
+import './SprintStatus.css';
 
 type SprintSegment = {
   label: string;
