+++ src/components/StatsCard.css
@@ -1,0 +1,64 @@
+.stats-card {
+  background: #fff;
+  border-radius: 20px;
+  padding: 1.25rem;
+  display: flex;
+  gap: 1rem;
+  align-items: center;
+  box-shadow: 0 10px 30px rgba(15, 23, 42, 0.08);
+}
+
+.stats-card__icon {
+  width: 48px;
+  height: 48px;
+  border-radius: 16px;
+  display: grid;
+  place-items: center;
+  font-size: 1.5rem;
+  background: #eff4ff;
+}
+
+.stats-card__content {
+  display: flex;
+  flex-direction: column;
+  gap: 0.2rem;
+}
+
+.stats-card__title {
+  margin: 0;
+  font-size: 0.95rem;
+  color: #64748b;
+  font-weight: 600;
+}
+
+.stats-card__value {
+  margin: 0;
+  font-size: 1.8rem;
+  font-weight: 700;
+}
+
+.stats-card__subtitle {
+  margin: 0;
+  font-size: 0.85rem;
+  color: #94a3b8;
+}
+
+.stats-card--default .stats-card__icon {
+  background: rgba(99, 102, 241, 0.1);
+}
+
+.stats-card--success .stats-card__icon {
+  background: rgba(52, 211, 153, 0.15);
+}
+
+.stats-card--danger .stats-card__icon {
+  background: rgba(248, 113, 113, 0.15);
+}
+
+.stats-card--warning .stats-card__icon {
+  background: rgba(251, 191, 36, 0.15);
+}
+
+.stats-card--info .stats-card__icon {
+  background: rgba(96, 165, 250, 0.15);
+}
