+++ src/components/TeamProgress.css
@@ -1,0 +1,47 @@
+.team-progress {
+  list-style: none;
+  margin: 0;
+  padding: 0;
+  display: flex;
+  flex-direction: column;
+  gap: 1.25rem;
+}
+
+.team-progress__item {
+  display: flex;
+  flex-direction: column;
+  gap: 0.75rem;
+}
+
+.team-progress__info {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 1rem;
+}
+
+.team-progress__name {
+  margin: 0;
+  font-weight: 600;
+  color: #0f172a;
+}
+
+.team-progress__value {
+  margin: 0;
+  font-size: 0.9rem;
+  color: #64748b;
+}
+
+.team-progress__bar {
+  width: 100%;
+  height: 14px;
+  background: #e2e8f0;
+  border-radius: 999px;
+  overflow: hidden;
+}
+
+.team-progress__fill {
+  height: 100%;
+  background: linear-gradient(90deg, #22c55e, #16a34a);
+  border-radius: 999px;
+}
