--- src/css/project_dashboard.css
@@ -1,13 +1,0 @@
-::-webkit-scrollbar {
-    display: none;
-}
-
-body {
-    font-family: 'Inter', sans-serif;
-}
-
-@media (min-width: 1024px) {
-    #chatbox-panel {
-        max-width: 24rem;
-    }
-}
